requirements.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
coverage==7.13.1
cymem==2.0.13
DAWG2-Python==0.9.0
distro==1.9.0
et_xmlfile==2.0.0
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
Levenshtein==0.27.3
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
numpy==2.4.1
openai==2.15.0
openpyxl==3.1.5
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pluggy==1.6.0
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymorphy3==2.0.6
pymorphy3-dicts-ru==2.4.417150.4580142
pymorphy3-dicts-uk==2.4.1.1.1663094765
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytz==2025.2
RapidFuzz==3.14.3
reportlab==4.4.7
requests==2.32.5
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.3.10
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uk_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.8.0/uk_core_news_sm-3.8.0-py3-none-any.whl#sha256=d20adb50b42c0dcfdedf4994dabcb96789a64983a9ab560d0c6c38a59e8efb58
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
xlrd==2.0.2
xmltodict==1.0.2
